FusionSpecies

Documentation for FusionSpecies.

FusionSpecies.get_species

FusionSpecies.get_species

FusionSpecies.get_species_Z

FusionSpecies.get_species_index

FusionSpecies.get_species_mass

FusionSpecies.get_atoms

FusionSpecies.get_atoms_index

FusionSpecies.get_electron

FusionSpecies.get_electron_index

FusionSpecies.get_imp_atoms

FusionSpecies.get_imp_atoms_index

FusionSpecies.get_imp_ions

FusionSpecies.get_imp_ions_index

FusionSpecies.get_ions

FusionSpecies.get_ions_index

FusionSpecies.get_main_atom

FusionSpecies.get_main_atom_index

FusionSpecies.get_main_ion

FusionSpecies.get_main_ion_index

FusionSpecies.get_molecules

FusionSpecies.get_molecules_index

FusionSpecies.get_neutrals

FusionSpecies.get_neutrals_index

FusionSpecies.get_species

FusionSpecies.get_species

FusionSpecies.show_elements

FusionSpecies.get_species — Method

get_species(element::FusionSpecies.Element) -> Any

return the species that are of type element

FusionSpecies.get_species — Method

get_species(

) -> Vector{<:FusionSpecies.AbstractLoadedSpecies}

get a list of active species

FusionSpecies.get_species_Z — Method

get_species_Z() -> Any

get a list of the charge state of active species

FusionSpecies.get_species_index — Method

get_species_index(
    s::FusionSpecies.AbstractLoadedSpecies
) -> Any

get a list of the index of active species

get_species_index(s)

defined at /Users/jeromeguterl/Dropbox/StataProject/StataMater/FusionSpecies.jl/src/getter.jl:208.

FusionSpecies.get_species_mass — Method

get_species_mass() -> Vector{Float64}

get a list of the mass of active species

FusionSpecies.get_atoms — Method

get_atoms() -> Vector{<:FusionSpecies.AbstractLoadedSpecies}

get a list of atoms among active species

FusionSpecies.get_atoms_index — Method

get_atoms_index() -> Vector{Int64}

get a list of index of atoms among active species

FusionSpecies.get_electron — Method

get_electron(; enforce) -> Any

return the electron species in active species

FusionSpecies.get_electron_index — Method

get_electron_index(; enforce) -> Union{Nothing, Int64}

return the index of the electron species

FusionSpecies.get_imp_atoms — Method

get_imp_atoms(

) -> Vector{<:FusionSpecies.AbstractLoadedSpecies}

get a list of ions among active species

FusionSpecies.get_imp_atoms_index — Method

get_imp_atoms_index() -> Vector{Int64}

get a list of index of impurity atoms among active species

FusionSpecies.get_imp_ions — Method

get_imp_ions(

) -> Vector{<:FusionSpecies.AbstractLoadedSpecies}

get a list of ions among active species

FusionSpecies.get_imp_ions_index — Method

get_imp_ions_index() -> Vector{Int64}

get a list of index of ions among active species

FusionSpecies.get_ions — Method

get_ions() -> Vector{<:FusionSpecies.LoadedSpecies}

get a list of ions among active species

FusionSpecies.get_ions_index — Method

get_ions_index() -> Vector{Int64}

get a list of index of ions among active species

FusionSpecies.get_main_atom — Method

get_main_atom(; enforce) -> Any

return the main ion species in active species

FusionSpecies.get_main_atom_index — Method

get_main_atom_index(; kw...) -> Int64

get the index of main atom among active species

FusionSpecies.get_main_ion — Method

get_main_ion(; enforce) -> Any

return the main ion species in active species

FusionSpecies.get_main_ion_index — Method

get_main_ion_index(; kw...) -> Int64

get a list of index of ions among active species

FusionSpecies.get_molecules — Method

get_molecules(

) -> Vector{<:FusionSpecies.AbstractLoadedSpecies}

get a list of molecules among active species

FusionSpecies.get_molecules_index — Method

get_molecules_index() -> Vector{Int64}

get a list of molecules among active species

FusionSpecies.get_neutrals — Method

get_neutrals() -> Any

get a list of atoms among active species

FusionSpecies.get_neutrals_index — Method

get_neutrals_index() -> Vector{Int64}

get a list of index of neutrals among active species

FusionSpecies.get_species — Method

get_species(species_index::Int64) -> Any

return the species with index species_index

FusionSpecies.get_species — Method

get_species(symbol::Symbol; lock) -> Any

return the species that are of type element

FusionSpecies.show_elements — Method

show_elements()

display available elements